requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.0.8
django-appconf==1.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-debug-toolbar==2.2
django-extensions==3.1.1
django-filter==2.3.0
django-select2==7.4.2
djangorestframework==3.11.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isodate==0.6.0
jdcal==1.4.1
openpyxl==3.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.8
pyDes==2.0.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
xlrd==1.2.0
zeep==3.4.0
zope.interface==5.2.0
